[refactor]: move domain entities into `domain` dir
