+++ src/core/domain/liquidity.ts
@@ -2,7 +2,7 @@ import type { Address, Deadline } from '../types'
 import { NATIVE_TOKEN, isNativeToken } from '../const'
 import { buildPair, buildPairAsync, TokensPair } from '@/utils/pair'
 import { computeTransactionFee, deadlineFiveMinutesFromNow } from '../utils'
-import Wei from './Wei'
+import { Wei } from '../entities'
 import { Agent, AgentPure } from './agent'
 import { TokensPure } from './tokens'
 import CommonContracts from './CommonContracts'
