+++ src/core/entities/index.ts
@@ -1,8 +1,1 @@
 export { default as Wei, type WeiAsToken, type WeiRaw } from './Wei'
-export { default as CommonContracts } from './CommonContracts'
-export { Agent, AgentPure } from './agent'
-export { Tokens, TokensPure } from './tokens'
-export { Swap, SwapPure } from './swap'
-export { Liquidity, LiquidityPure } from './liquidity'
-export { Farming, type PoolId, EarnPure, Earn, Staking } from './earn'
-export { Dex, DexPure } from './dex'
