+++ src/core/index.ts
@@ -1,6 +1,7 @@
 export * from './types'
 export * from './const'
 export * from './entities'
+export * from './domain'
 export * from './use-web3'
 export { isAddress, formatAddress, isEmptyAddress, parseAddress, deadlineFiveMinutesFromNow } from './utils'
 export { AbiLoader } from './abi'
